requirements.txt
streamlit
langchain
faiss-cpu
sentence-transformers
PyPDF2
python-docx
ollama
